update sparql queries - update prefixes - refactor keyword queries
