+++ code/graph-connector/app/rdf/document_queries.py
@@ -1,4 +1,4 @@
-from .common import SG_PREFIXES, PREFIXES
+from .common import PREFIXES
 
 
 def get_document_name_list_query(keys: str = '*', limit: int = 10, page: int = 1) -> str:
@@ -28,16 +28,16 @@ WHERE {{
         dc:issued ?issued ;
         dc:creator ?author ;
         dc:abstract ?abstract ;
-        sgp:website ?website;
+        dgp:website ?website;
         datacite:doi ?doi ;
         dc:language ?language .
         OPTIONAL {{
-            <{publication_uri}> sgp:attribute ?atri .
-            ?atri sgp:status ?atri_s;
+            <{publication_uri}> dgp:attribute ?atri .
+            ?atri dgp:status ?atri_s;
             rdfs:subClassOf ?atr .
             ?atr rdf:value ?atr_v ;
                 rdfs:subClassOf ?atrf .
-            ?atrf sgp:name ?atr_n .
+            ?atrf dgp:name ?atr_n .
             BIND(CONCAT('{{"name":"',?atr_n,'","value":"',?atr_v,'","status":',STR(?atri_s),'}}') as ?att) .
         }}
 }}
