+++ code/graph-connector/app/rdf/queries.py
@@ -15,11 +15,11 @@ def get_keyword_cross_reference_query(keywords: list[dict], limit: int = default
 SELECT ?cross_keyword ?cross_value (COUNT(?kwic) AS ?occurrences)
 WHERE {{
     ?pub rdf:type foaf:Document .
-    ?pub sgp:keyword ?kwi .
+    ?pub dgp:keyword ?kwi .
     ?kwi rdf:type ?kw .
     ?kw rdf:value ?val 
     FILTER(?val IN {keyword_list_string}) .
-    ?pub sgp:keyword ?kwic .
+    ?pub dgp:keyword ?kwic .
     ?kwic rdf:type ?cross_keyword .
     ?cross_keyword rdf:value ?cross_value 
     FILTER(?cross_value NOT IN {keyword_list_string}) .
@@ -34,28 +34,6 @@ LIMIT {limit}
 """
 
 
-def get_keyword_begins_with_query(begins_with: str, limit: int = default_limit,
-                                  attributes: list[AdditionalAttribute] = None,
-                                  years_span: tuple[int, int] = None) -> str:
-    if limit is None:
-        limit = default_limit
-    return f"""
-{PREFIXES}
-
-SELECT DISTINCT ?keyword_value
-WHERE {{
-    ?pub rdf:type foaf:Document .
-    ?pub sgp:keyword ?kwi .
-    ?kwi rdf:type ?keyword .
-    ?keyword rdf:value ?keyword_value
-    FILTER regex(?keyword_value, "^{begins_with}", "i") .
-    {get_attributes_filter(attributes)}
-    {get_release_year_filter(years_span)}
-}}
-LIMIT {limit}
-"""
-
-
 def get_publications_result(keywords: list[dict], limit: int = default_limit,
                             attributes: list[AdditionalAttribute] = None,
                             years_span: tuple[int, int] = None, page: int = 1) -> str:
@@ -66,7 +44,7 @@ SELECT ?pub ?title ?issued ?doi ?language (concat('["',GROUP_CONCAT(?author;sepa
 (COUNT(?kwi) AS ?occurrences)
 WHERE {{
     ?pub rdf:type foaf:Document .
-    ?pub sgp:keyword ?kwi .
+    ?pub dgp:keyword ?kwi .
     ?kwi rdf:type ?kw .
     ?kw rdf:value ?val 
     FILTER(?val in {get_keyword_list_string(keywords)}) .
@@ -87,31 +65,14 @@ OFFSET {(page - 1 ) * limit}
 """
 
 
-def get_keywords_query(publication_uri: str, limit: int = default_limit, page: int = 1) -> str:
-    return f"""
-{SG_PREFIXES}
-
-SELECT ?keyword_value ?verification_status ?kw
-WHERE {{
-    <{publication_uri}> sgp:keyword ?kwi .
-    ?kwi rdf:type ?kw ;
-        sgp:status ?verification_status .
-    ?kw rdf:value ?keyword_value .
-}}
-
-LIMIT {limit}
-OFFSET {(page - 1 ) * limit}
-"""
-
-
 def get_attributes_names_query() -> str:
     return f"""
 {SG_PREFIXES}
 
 SELECT ?attr_name
 WHERE {{
-    ?attr_flavour rdf:type sgc:Attribute .
-    ?attr_flavour sgp:name ?attr_name .
+    ?attr_flavour rdf:type dgc:Attribute .
+    ?attr_flavour dgp:name ?attr_name .
 }}
 """
 
@@ -121,8 +82,8 @@ def get_attribute_values_query(attribute_name:str) -> str:
 
 SELECT ?attr_value
 WHERE {{
-    ?attr_flavour rdf:type sgc:Attribute .
-    ?attr_flavour sgp:name "{attribute_name}" .
+    ?attr_flavour rdf:type dgc:Attribute .
+    ?attr_flavour dgp:name "{attribute_name}" .
     ?attr rdfs:subClassOf ?attr_flavour .
     ?attr rdf:value ?attr_value .
 }}
